+++ met-api/tests/unit/services/test_comment.py
@@ -17,7 +17,6 @@
 Test-Suite to ensure that the Comment service routines are working as expected.
 """
 from met_api.services.comment_service import CommentService
-
 from tests.utilities.factory_utils import (
     factory_comment_model, factory_submission_model, factory_survey_and_eng_model, factory_user_model)
 
@@ -31,15 +30,3 @@ def test_get_comments(session):  # pylint:disable=unused-argument
     comment_records = CommentService().get_comments_by_submission(submission.id)
     assert len(comment_records) == 1
     assert comment_records[0]['status_id'] == 1
-
-
-def test_review_comment(client, jwt, session):  # pylint:disable=unused-argument
-    """Assert that a comment can be reviewed."""
-    admin_user = factory_user_model(3)
-    user_details = factory_user_model()
-    survey, eng = factory_survey_and_eng_model()
-    submission = factory_submission_model(survey.id, user_details.id)
-    factory_comment_model(survey.id, submission.id)
-    comment_records = CommentService().review_comment(submission.id, 2, admin_user.external_id)
-    assert len(comment_records) == 1
-    assert comment_records[0]['status_id'] == 2
